fix set pct
Former-commit-id: a8290aebd8c5369ac88c7c6e43898b5fb2cc5be9

+++ src/services/rpc/wallet/assetwalletrpc.cpp
@@ -38,7 +38,7 @@ CAmount getAuxFee(const CAuxFeeDetails &auxFeeDetails, const CAmount& nAmount) {
         nBoundAmount = fee.nBound;
         nNextBoundAmount = feeNext.nBound;
         // max uint16 (65535 = 0.65535 = 65.5535%)
-        nRate = fee.nPercent / 100000;
+        nRate = ((double)fee.nPercent) / 100000.0;
         // case where amount is in between the bounds
         if(nAmount >= nBoundAmount && nAmount < nNextBoundAmount){
             break;
